End of Day 9?
Very good day, made several great improvements to the app. Next step is to get the markers to instead be circles coloured according to a colour grouping based on risk. Then get the table and map talking to each other so that if the table is clicked on or filtered the map responds.